Changed runlen to a matrix with 1st entry indicating length of run in meters and second entry indicating number of stops in run, getTrainingSamples was also modified to use the ping data directly in extractAC instead of using Dr. Kargls acolor.exe output i.e. ob.acp

+++ extractAC.m
@@ -22,7 +22,7 @@ AC  = AC(:,sum(abs(AC),1)>1);
 
 % Decimate along aspect
 for i = 1:size(AC,1)
-aAC(i,:) = resample(AC(i,:),aper,size(AC,2));
+aAC(i,:) = resample(double(AC(i,:)),aper,size(AC,2));
 end
 % Select enhanced aspects in the in center (the 'norm' ones all have this)
 aAC = aAC(:,sum(abs(aAC),1)>eps);
